move tests to seperate folder
